winrate-honest-denominator: A — split analyze_basic.py under the file-size limit
Extract methodsUsed metadata and pure aggregation helpers out of
analyze_basic.py into analyze_basic_metadata.py and
analyze_basic_aggregation.py so the entry-point script lands under the
400-line file-size limit (was 704 lines, now 359 lines). Zero behavior
change — test_analyze_basic.py passes with no test modifications.
